requirements.txt
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncgTTS==0.3.1
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
discord.py==2.0.0
flake8==4.0.1
frozenlist==1.3.0
gTTS==2.2.4
idna==3.3
lxml==4.9.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
